backend/requirements.txt
flask
flask-cors
pillow
pytesseract
openai